Test case fixes needed for modified rule syntax.
Because it no longer be assumed that all SPARQL functions and operators
occur in a `filter` block, `filter` must now be explicitly specified.

The change of default select-type default has been reverted, as this
breaks the standard query pattern, and is in fact not needed
(cf. https://github.com/UCDenver-ccp/kr/pull/4).