Add support for operator validation.
It may be desirable in some instances to pre-validate the arguments that
are passed to an operator, rather than only relying on the backend
parsing of the generated SPARQL.

One example of this arises from how operators are implemented across
different backends.  For example, the IRI function across the various
platforms does not consistent handle language-tagged strings; some (like
Sesame) will do the conversion, while others (like Allegrograph) will
not.

This also provides a hook to insert validatation to help users with the
not-quite-SPARQL pattern syntax.